Add Params module based on the one in gnome-shell
This modules offers a convenient way to pass arguments to complex
functions, in a way similar to keyword arguments in Python.
To make it also suited for useful for inheriting from GObject classes,
as the convention is to use a property bag in the constructor,
a new function is introduced, filter(), which allows extending
the set of supported properties in a trasparent way.
Also, the confusing allowExtras boolean parameter is removed,
in favor of two functions, parse() which throws and fill() which
doesn't.
With lots of documentation, code examples and tests.